CC: Fix more bugs
Some type conversion problems

+++ vars/chaincode/elec-chaincode/java/src/main/java/org/hyperledger/fabric/chaincode/Account.java
@@ -17,11 +17,11 @@ public class Account {
         return accountId;
     }
 
-    public Double getElecAmount() {
+    public int getElecAmount() {
         return elecAmount;
     }
 
-    public Double getBalance() {
+    public int getBalance() {
         return balance;
     }
